[#390] Typography Symbols are not mapped to universal keyboard standards.
Replaces specific typography characters with universal standards.
This is available to the user by a default checked checkbox.

The option is given to the user rather than always automatically applied because:
a) I'm not 100% sure other languages or keyboard layouts do not use these typography symbols as standard inputs.
b) There may be some who enjoy the challenge of entering Shift+Option+[ just to get a fancy quote mark.