Opposite Shift, update key list to add symbols (#1)
* update key list to add symbols

Added the symbols that are on the right side of the keyboard and the symbols that are on the left side to their appropriate lists. Left the `^` key because it is neutral.

* Edit opposite shift text

Not all languages use shift for capitalization, so with the added symbols, it doesn't make a lot of sense.