Added some quotes and fixed a typo. (#1730)
* Update turkish.json

There is a typo, the word "tabi" is even has a meaning too but it's not common used in Turkish language and probably it meant to be "tabii"(means of course). I don't know is this even a problem but i wanted to notify you about that.

* Update turkish.json

Added some quotes.