Add Japanese as a language to Monkeytype (#1919) by sashinexists
* Added Japanese to _groups file

* Added Japanese Hiragana and Katakana to _list.json

* Create japanese_hiragana.json

A list of the most common hiragana only words. That it's only hiragana means all of Monkeytype's features should work just fine.
Tested with custom input on the Mozc and KKC keyboard layouts. You have to press enter to bypass Kanji conversion.

* Create japanese_katakana.json

A list of loan words commonly used in Japan. Everything is in Katakana. This should work with the "Japanese (kana)" keyboard layout.